Move getCity logic to API folder
